Adds a number of features:
There is now a Options tab.  Currently, this allows you to:

* Show long or short time on the chat page
* Provide protection for ban/unban actions
* Alter if you wish to get alerted for Community appeals.
* Change how many common accounts to find

Speaking of appeal alerts, this has those now.  When you first get the update, your browser will ask you if you would like notifications.
Should you say yes, then neat little desktop-style notifications will be displayed for you when someone appeals one of your bans, or when they respond.
If you check the box on the options page, you will also be alerted of Community appeals/responses.

Another feature this release provides is Fingerprint-similarities.

On a profile page, you will see a button below the fingerprints for a user.  If you click the button, the script will churn through all of their fingerprints, and return the accounts that showed up the most on them.  This can be useful for confirming ban evasion.  Be warned though that it can take a little while to run and is a little resource heavy, so use it sparingly.

Other than that, there was some formatting standardization which is always nice.