New functionality and sped up fingerprint checker
Added coded to highlight high risk IPs as identified by the ban evasion log in the mod subreddit, IPs will be scanned against known bad IPs and the portions of the IP that match from the beginning will be highlighted in red and underscored.  The IPs are stored in a separate included file so that they may be updated without requiring an increment update to the main script.  

Increased the scanning speed of the fingerprint checker significantly by utilizing ajax calls and setTimeout to create pseudo threads up to 10 at a time checking the fingerprints.

Also increased version to 1.3 as this contains new feature(s).