Mass Ban Evasion
Adding a ban account button.  Clicking this will prompt TWO confirmations that make sure you want to do this.  When you confirm, it will add a ban to every IP and every account under that Ban Evasion Profile and refresh the page.  As a side effect of this implementation, IP's in the Ban Evasion section will also tell you if they are currently banned or not.