Update script to work with support site refresh
A lot of the workaround that existed could be pulled out which was nice. We also no longer need to AJAX out to the main site to grab user information since it’s now part of the ticketing system, which is also nice.

This returns to basically just being a canned reply generator with some WYSYWIG capabilities.